Updating fork

+++ src/tileray.cpp
@@ -3,7 +3,7 @@
 #include <cmath>
 #include <cstdlib>
 
-#include "game_constants.h"
+#include "math_defines.h"
 
 static const int sx[4] = { 1, -1, -1, 1 };
 static const int sy[4] = { 1, 1, -1, -1 };
@@ -111,7 +111,7 @@ int tileray::dir8() const
 
 // This function assumes a vehicle is being drawn.
 // It assumes horizontal lines are never skewed, vertical lines often skewed.
-long tileray::dir_symbol( long sym ) const
+int tileray::dir_symbol( int sym ) const
 {
     switch( sym ) {
         // output.cpp special_symbol() converts yubn to corners, hj to lines, c to cross
